% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPHIGlobalComparisions-package.R
\docType{package}
\name{OPHIDataBankGlobalComparison-package}
\alias{OPHIDataBankGlobalComparison}
\alias{OPHIDataBankGlobalComparison-package}
\title{OPHIDataBankGlobalComparison: A golem package to create an app for global comparisions of Multidimensional Poverty Index}
\description{
OPHIDataBankGlobalComparison package allows users to acces a web application through which they can interact with the OPHI data bank. The users will be able to see comparisions of Aggregate measures (MPI,H,A), comparision of Headcount of poverty (MPI vs $1.90/day) and contribution of various indicators to poverty.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AyushBipinPatel/OPHIDataBankGlobalComparison}
  \item Report bugs at \url{https://github.com/AyushBipinPatel/OPHIDataBankGlobalComparison/issues}
}

}
\author{
\strong{Maintainer}: Ayush Patel \email{ayush.ap58@gmail.com}

}
\keyword{internal}
